[Tests-only] Refactor Traits and context files in acceptance tests
